+++ R/highlight.R
@@ -75,6 +75,11 @@ highlight <- function(text, classes = classes_chroma(), pre_class = NULL, code =
   new <- style_token(changes$escaped, changes$href, changes$class)
   out <- replace_in_place(parsed$text, start, end, replacement = new)
 
+  # Add per-line span to match pandoc
+  out <- strsplit(out, "\n")[[1]]
+  out <- paste0("<span>", out, "</span>")
+  out <- paste0(out, collapse = "\n")
+
   if (!is.null(pre_class)) {
     out <- paste0(
       "<pre class='", paste0(pre_class, collapse = " "), "'>\n",